% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera_plots.R
\name{camera_sankey}
\alias{camera_sankey}
\title{Sankey diagram of CAMERA annotation results}
\usage{
camera_sankey(cam_result, height = 800, margin_right = 160)
}
\arguments{
\item{cam_result}{An \code{xsAnnotate} object returned by
\code{CAMERA::findAdducts()}. The rules used during annotation are read
from \code{cam_result@ruleset}.}

\item{height}{Plot height in pixels. Default \code{800}.}

\item{margin_right}{Right margin in pixels, to leave room for the rightmost
node labels. Default \code{160}.}
}
\value{
A \code{plotly} htmlwidget.
}
\description{
Draws an interactive four-column Sankey diagram showing how peaks annotated
by CAMERA flow from isotope type through multiplicity to adduct rule.
}
\seealso{
\code{\link{camera_sankey_data}}
}
\author{
Jan Stanstrup, \email{stanstrup@gmail.com}
}
