% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{MZ_CAMERA}
\alias{MZ_CAMERA}
\title{Load adduct/fragment rules for CAMERA}
\usage{
MZ_CAMERA(mode, warn_clash = TRUE, clash_ppm)
}
\arguments{
\item{mode}{The ionization mode. Can be "pos", "neg" or "ei".}

\item{warn_clash}{Warn for adducts/fragments that gives rise to the same m/z difference in spectra.}

\item{clash_ppm}{ppm to use for the above check.}
}
\value{
tibble defining adduct rules.
}
\description{
This function loads rules for use with CAMERA adduct annotation.
}
\examples{
MZ_CAMERA(mode="pos", warn_clash = TRUE, clash_ppm=30)


}
\author{
Jan Stanstrup, \email{stanstrup@gmail.com}
}
